Create CountTranscript
This program was written to evaluate participation in an online chat. Program needs reference list of participants. Program reads the chat transcript and does a word count for each participants. Once word count is done, it scores the participants based on a median value. All above median get 1, below get 0